Move DomainRegistrationDetailsViewModel to ui.domains package
